# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.3.0
  description: Front Office API — league management, results, standings, and scheduling.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leagues:
    post:
      tags: [Leagues]
      operationId: createLeague
      summary: Create a new league
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLeagueInput'
      responses:
        '201':
          description: League created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/League'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Slug already taken.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  /leagues/{leagueId}:
    parameters:
      - $ref: '#/components/parameters/LeagueId'
    get:
      tags: [Leagues]
      operationId: getLeague
      summary: Get a league
      responses:
        '200':
          description: League found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/League'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    patch:
      tags: [Leagues]
      operationId: updateLeague
      summary: Update league presentation metadata (commissioner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLeagueInput'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/League'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  /leagues/{leagueId}/listing:
    parameters:
      - $ref: '#/components/parameters/LeagueId'
    get:
      tags: [Discovery]
      operationId: getLeagueListing
      summary: Get league listing / discovery settings (commissioner only)
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Listing settings for this league.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeagueListing'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Forbidden — caller is not the league commissioner.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: League not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    patch:
      tags: [Discovery]
      operationId: updateLeagueListing
      summary: Update league listing / discovery settings (commissioner only)
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLeagueListingInput'
      responses:
        '200':
          description: Updated listing settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeagueListing'
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Forbidden — caller is not the league commissioner.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: League not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  /leagues/{leagueId}/seasons:
    parameters:
      - $ref: '#/components/parameters/LeagueId'
    get:
      tags: [Leagues]
      operationId: listSeasons
      summary: List seasons for a league
      responses:
        '200':
          description: Seasons ordered oldest first.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Season'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      tags: [Leagues]
      operationId: createSeason
      summary: Create a new season and provision the configured number of franchise seats (commissioner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSeasonInput'
      responses:
        '201':
          description: Season created with seats from the active league settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Season'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'

  # ------------------------------------------------------------------ Seats
...
components:
  parameters:
...
    LeagueId:
      name: leagueId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 problem details.
      required: [type, title, status, trace_id]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: ["string", "null"]
        instance:
          type: ["string", "null"]
        trace_id:
          type: string
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ProblemError'
    ProblemError:
      type: object
      properties:
        pointer:
          type: string
        message:
          type: string
...
    League:
      type: object
      required: [id, slug, name, visibility]
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        visibility:
          type: string
          enum: [public, unlisted, private]
        logo_url:
          type: ["string", "null"]
        primary_color:
          type: ["string", "null"]
        secondary_color:
          type: ["string", "null"]
        owner_user_id:
          type: string
        created_at:
          type: string
          format: date-time
    CreateLeagueInput:
      type: object
      required: [name, slug, settings]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        slug:
          type: string
          minLength: 2
          maxLength: 40
          pattern: '^[a-z0-9][a-z0-9-]*[a-z0-9]$'
        visibility:
          type: string
          enum: [public, unlisted, private]
          default: public
        primary_color:
          type: ["string", "null"]
        secondary_color:
          type: ["string", "null"]
        logo_url:
          type: ["string", "null"]
        settings:
          $ref: '#/components/schemas/LeagueSettingsInput'
    UpdateLeagueInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        visibility:
          type: string
          enum: [public, unlisted, private]
        primary_color:
          type: ["string", "null"]
        secondary_color:
          type: ["string", "null"]
        logo_url:
          type: ["string", "null"]
...
    CreateSeasonInput:
      type: object
      required: [label, game_title]
      properties:
        label:
          type: string
          minLength: 1
          maxLength: 100
        game_title:
          type: string
          minLength: 1
          maxLength: 50
        starts_on:
          type: ["string", "null"]
          format: date
        ends_on:
          type: ["string", "null"]
          format: date
        salary_cap_cents:
          type: ["integer", "null"]
          minimum: 0
        roster_min:
          type: ["integer", "null"]
          minimum: 1
        roster_max:
          type: ["integer", "null"]
          minimum: 1
        games_per_matchup:
          type: integer
          minimum: 1
          maximum: 8
        points_win:
          type: integer
          minimum: 0
        points_ot_loss:
          type: integer
          minimum: 0
        points_reg_loss:
          type: integer
          minimum: 0
        tiebreakers:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
...
    Season:
      type: object
      required: [id, ordinal, label, game_title]
      properties:
        id:
          type: string
        league_id:
          type: string
        settings_version_id:
          type: ["string", "null"]
          description: Immutable league settings version bound to this season.
        ordinal:
          type: integer
          minimum: 1
        label:
          type: string
        game_title:
          type: string
        starts_on:
          type: ["string", "null"]
        ends_on:
          type: ["string", "null"]
        salary_cap_cents:
          type: ["integer", "null"]
        roster_min:
          type: ["integer", "null"]
        roster_max:
          type: ["integer", "null"]
        games_per_matchup:
          type: integer
        points_win:
          type: integer
        points_ot_loss:
          type: integer
        points_reg_loss:
          type: integer
        tiebreakers:
          type: array
          items:
            type: string
        is_active:
          type: boolean

    # ------------------------------------------------------------------ Seats
...
    LeagueListing:
      type: object
      required: [league_id, is_listed, accepting_signups, accepting_waitlist]
      properties:
        league_id:
          type: string
        is_listed:
          type: boolean
        accepting_signups:
          type: boolean
        accepting_waitlist:
          type: boolean
        blurb:
          type: ['string', 'null']
        platform:
          type: ['string', 'null']
          enum: [psn, playstation, xbox, both, crossplay, null]
        competitiveness:
          type: ['string', 'null']
          enum: [casual, competitive, hardcore, null]
        suggested_division:
          type: ['string', 'null']
        timezone_focus:
          type: ['string', 'null']
        listed_at:
          type: ['string', 'null']
          format: date-time
        updated_at:
          type: ['string', 'null']
          format: date-time
    UpdateLeagueListingInput:
      type: object
      properties:
        is_listed:
          type: boolean
        accepting_signups:
          type: boolean
        accepting_waitlist:
          type: boolean
        blurb:
          type: ['string', 'null']
        platform:
          type: ['string', 'null']
          description: Canonical values plus legacy aliases accepted during client migration.
          enum: [psn, playstation, xbox, both, crossplay, null]
        competitiveness:
          type: ['string', 'null']
          enum: [casual, competitive, hardcore, null]
        suggested_division:
          type: ['string', 'null']
        timezone_focus:
          type: ['string', 'null']
...